Build: Use vendored dependencies for building .deb packages
It is currently not possible to build Squeekboard with only
packages that are available in Debian.
This is not likely to change until Squeekboard has been ported
to GTK4 or gtk4-layer-shell.
Until then, vendor the dependencies during the build,
so that one can still use the ".deb"-packages for testing.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/659>