Cargo.toml: Simplify versions
This should not change the selection of the versions to update to.

Cargo considers the leftmost non-zero part of the version, to be the
number that changes compatibility, and will not update to a version that
it considers to be incompatible.

For more information, read:
https://doc.rust-lang.org/cargo/reference/resolver.html#semver-compatibility

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/676>