layouts: Adjust layout-sizes to the correct aspect-ratios
and fix some other issues, to make the layouts more consistent.

The aspect-ratios of the layouts,
should now match the intended aspect-ratios (12:7 and 16:5),
with 3 decimal-digits of precision; either exactly,
or rounded to be a little more square,
so that the entire height of the panel will be used.
Adding 0.001 to `width` in the `default`-outlines of the layouts,
would make the layouts a little wider than the intended aspect-ratio.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/691>