keyboard.rs: Add keycode to the keymap with BackSpace if there is no other
BackSpace does not work with `tools/entry.py` (made with GTK3),
if the keymap with BackSpace does not contain any other keycodes.
This should only happen for the emoji-layout or incomplete custom-layouts,
because the layout-tests for normal layouts
check for the presence of a button for Return.
This does add an "Unknown"-keycode, if necessary,
to let BackSpace work anyway.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/695>