parsing: fix test for new xkbcommon version
The latest version of `xkbcommon` (0.8) fixes a bug where
`keysym_to_utf8` returned a string including the terminating NULL
character. The `unicode_keysym` test used to rely on this behaviour, so
we must update it so it still works with the new version.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/697>