CI: Remove 'check_release'-CI-job and 'check_release.py'
This CI-job is a reminder to publish a tag,
when there is no tag for the current version-number.

This does not work well when the version-number is increased
after a release, because there will be no such tag,
until the next release has been published.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/701>